+++ src/components/button/Button.tsx
@@ -1,0 +1,11 @@
+interface ButtonProps {
+  label: string;
+}
+
+export const Button = ({ label }: ButtonProps) => {
+  return (
+    <button>
+      {label}
+    </button>
+  );
+};
